+++ AlphaFS.UnitTest/File Class/ExportImportEncryptedFileRaw.cs
@@ -31,7 +31,7 @@ namespace AlphaFS.UnitTest
       // Pattern: <class>_<function>_<scenario>_<expected result>
 
       [TestMethod]
-      public void File_ExportImportEncryptedFileRaw_RountripOfFile_FileContentsOfImportedFileMatchesTheOriginalFile_LocalUNC_Success()
+      public void File_ExportImportEncryptedFileRaw_RountripOfFile_FileContentsOfImportedFileMatchesTheOriginalFile_LocalAndUNC_Success()
       {
          File_ExportImportEncryptedFileRaw(false);
          File_ExportImportEncryptedFileRaw(true);