+++ CHANGELOG.md
@@ -65,6 +65,7 @@ Version 2.1
 - Issue #201: Some exceptions contain an incorrect `HRESULT` (Thomas Levesque)
 - Issue #203: Directory.GetDirectories and GetFiles return absolute paths when given relative argument.
 - Issue #204: Giving empty string to GetFileName and related methods throws exception.
+- Issue #206: GetLastWriteTime throws exception for non-existing path.
 - Issue #217: Find.Replace() raises an exception.
 - Issue #219: Mismatching Implementation to System.IO.Path.GetDirectoryName.
 - Issue #226: DirectoryInfo using searchoption.